feat: add unit tests and move existing
